Updated Homescreen, Upload and ROI Panels
Homescreen now contains a three image slide show that utilizes threading to consistently move the images. It starts and stops when ever a user moves to another panel and returns to the home screen.

As for the ROI and Upload  panels, their panels that changed from fileUIController, are not synced and can changed when the other change if no files are present within the current upload actions.